Refactor air private functionality to respective builtin files
