+++ pkg/vm/builtins/bitwise.go
@@ -3,6 +3,8 @@ package builtins
 import (
 	"errors"
 	"fmt"
+	"math/big"
+	"sort"
 
 	"github.com/NethermindEth/cairo-vm-go/pkg/vm/memory"
 	"github.com/consensys/gnark-crypto/ecc/stark-curve/fp"
@@ -10,7 +12,7 @@ import (
 
 const BitwiseName = "bitwise"
 
-const CellsPerBitwise = 5
+const cellsPerBitwise = 5
 const inputCellsPerBitwise = 2
 const instancesPerComponentBitwise = 1
 
@@ -25,7 +27,7 @@ func (b *Bitwise) CheckWrite(
 }
 
 func (b *Bitwise) InferValue(segment *memory.Segment, offset uint64) error {
-	bitwiseIndex := offset % CellsPerBitwise
+	bitwiseIndex := offset % cellsPerBitwise
 	// input cell
 	if bitwiseIndex < inputCellsPerBitwise {
 		return errors.New("cannot infer value from input cell")
@@ -98,5 +100,52 @@ func (b *Bitwise) String() string {
 }
 
 func (b *Bitwise) GetAllocatedSize(segmentUsedSize uint64, vmCurrentStep uint64) (uint64, error) {
-	return getBuiltinAllocatedSize(segmentUsedSize, vmCurrentStep, b.ratio, inputCellsPerBitwise, instancesPerComponentBitwise, CellsPerBitwise)
+	return getBuiltinAllocatedSize(segmentUsedSize, vmCurrentStep, b.ratio, inputCellsPerBitwise, instancesPerComponentBitwise, cellsPerBitwise)
+}
+
+type AirPrivateBuiltinBitwise struct {
+	Index int    `json:"index"`
+	X     string `json:"x"`
+	Y     string `json:"y"`
+}
+
+func (b *Bitwise) GetAirPrivateInput(bitwiseSegment *memory.Segment) []AirPrivateBuiltinBitwise {
+	valueMapping := make(map[int]AirPrivateBuiltinBitwise)
+	for index, value := range bitwiseSegment.Data {
+		if !value.Known() {
+			continue
+		}
+		idx, typ := index/cellsPerBitwise, index%cellsPerBitwise
+		if typ >= 2 {
+			continue
+		}
+
+		builtinValue, exists := valueMapping[idx]
+		if !exists {
+			builtinValue = AirPrivateBuiltinBitwise{Index: idx}
+		}
+
+		valueBig := big.Int{}
+		value.Felt.BigInt(&valueBig)
+		valueHex := fmt.Sprintf("0x%x", &valueBig)
+		if typ == 0 {
+			builtinValue.X = valueHex
+		} else {
+			builtinValue.Y = valueHex
+		}
+		valueMapping[idx] = builtinValue
+	}
+
+	values := make([]AirPrivateBuiltinBitwise, 0)
+
+	sortedIndexes := make([]int, 0, len(valueMapping))
+	for index := range valueMapping {
+		sortedIndexes = append(sortedIndexes, index)
+	}
+	sort.Ints(sortedIndexes)
+	for _, index := range sortedIndexes {
+		value := valueMapping[index]
+		values = append(values, value)
+	}
+	return values
 }
