+++ pkg/vm/builtins/ecop.go
@@ -3,16 +3,19 @@ package builtins
 import (
 	"errors"
 	"fmt"
+	"math/big"
+	"sort"
 
 	"github.com/NethermindEth/cairo-vm-go/pkg/utils"
+	"github.com/NethermindEth/cairo-vm-go/pkg/vm/memory"
 	mem "github.com/NethermindEth/cairo-vm-go/pkg/vm/memory"
 	"github.com/consensys/gnark-crypto/ecc/stark-curve/fp"
 	"github.com/holiman/uint256"
 )
 
 const EcOpName = "ec_op"
-const CellsPerEcOp = 7
-const InputCellsPerEcOp = 5
+const cellsPerEcOp = 7
+const inputCellsPerEcOp = 5
 const instancesPerComponentEcOp = 1
 
 var feltThree fp.Element = fp.Element(
@@ -44,8 +47,8 @@ func (e *EcOp) InferValue(segment *mem.Segment, offset uint64) error {
 		return segment.Write(offset, &mv)
 	}
 	// get the current slot index and verify it is an output cell
-	ecopIndex := offset % CellsPerEcOp
-	if ecopIndex < InputCellsPerEcOp {
+	ecopIndex := offset % cellsPerEcOp
+	if ecopIndex < inputCellsPerEcOp {
 		return errors.New("cannot infer value from input cell")
 	}
 
@@ -99,7 +102,7 @@ func (e *EcOp) InferValue(segment *mem.Segment, offset uint64) error {
 	}
 
 	// store the resulting point `r`
-	outputOff := inputOff + InputCellsPerEcOp
+	outputOff := inputOff + inputCellsPerEcOp
 
 	// store the x and y coordinates of the resulting point
 	e.cache[outputOff] = r.X
@@ -111,7 +114,7 @@ func (e *EcOp) InferValue(segment *mem.Segment, offset uint64) error {
 }
 
 func (e *EcOp) GetAllocatedSize(segmentUsedSize uint64, vmCurrentStep uint64) (uint64, error) {
-	return getBuiltinAllocatedSize(segmentUsedSize, vmCurrentStep, e.ratio, InputCellsPerEcOp, instancesPerComponentEcOp, CellsPerEcOp)
+	return getBuiltinAllocatedSize(segmentUsedSize, vmCurrentStep, e.ratio, inputCellsPerEcOp, instancesPerComponentEcOp, cellsPerEcOp)
 }
 
 // structure to represent a point in the elliptic curve
@@ -233,4 +236,60 @@ func ecdouble(p *point, alpha *fp.Element) point {
 	y.Sub(&y, &p.Y)
 
 	return point{x, y}
+}
+
+type AirPrivateBuiltinEcOp struct {
+	Index int    `json:"index"`
+	PX    string `json:"p_x"`
+	PY    string `json:"p_y"`
+	M     string `json:"m"`
+	QX    string `json:"q_x"`
+	QY    string `json:"q_y"`
+}
+
+func (e *EcOp) GetAirPrivateInput(ecOpSegment *memory.Segment) []AirPrivateBuiltinEcOp {
+	valueMapping := make(map[int]AirPrivateBuiltinEcOp)
+	for index, value := range ecOpSegment.Data {
+		if !value.Known() {
+			continue
+		}
+		idx, typ := index/cellsPerEcOp, index%cellsPerEcOp
+		if typ >= inputCellsPerEcOp {
+			continue
+		}
+
+		builtinValue, exists := valueMapping[idx]
+		if !exists {
+			builtinValue = AirPrivateBuiltinEcOp{Index: idx}
+		}
+
+		valueBig := big.Int{}
+		value.Felt.BigInt(&valueBig)
+		valueHex := fmt.Sprintf("0x%x", &valueBig)
+		if typ == 0 {
+			builtinValue.PX = valueHex
+		} else if typ == 1 {
+			builtinValue.PY = valueHex
+		} else if typ == 2 {
+			builtinValue.QX = valueHex
+		} else if typ == 3 {
+			builtinValue.QY = valueHex
+		} else if typ == 4 {
+			builtinValue.M = valueHex
+		}
+		valueMapping[idx] = builtinValue
+	}
+
+	values := make([]AirPrivateBuiltinEcOp, 0)
+
+	sortedIndexes := make([]int, 0, len(valueMapping))
+	for index := range valueMapping {
+		sortedIndexes = append(sortedIndexes, index)
+	}
+	sort.Ints(sortedIndexes)
+	for _, index := range sortedIndexes {
+		value := valueMapping[index]
+		values = append(values, value)
+	}
+	return values
 }
