+++ pkg/vm/builtins/keccak.go
@@ -4,6 +4,8 @@ import (
 	"encoding/binary"
 	"errors"
 	"fmt"
+	"math/big"
+	"sort"
 
 	"github.com/NethermindEth/cairo-vm-go/pkg/vm/memory"
 	"github.com/consensys/gnark-crypto/ecc/stark-curve/fp"
@@ -18,8 +20,8 @@ import (
 //
 
 const KeccakName = "keccak"
-const CellsPerKeccak = 16
-const InputCellsPerKeccak = 8
+const cellsPerKeccak = 16
+const inputCellsPerKeccak = 8
 const instancesPerComponentKeccak = 16
 
 type Keccak struct {
@@ -37,14 +39,14 @@ func (k *Keccak) InferValue(segment *memory.Segment, offset uint64) error {
 		mv := memory.MemoryValueFromFieldElement(&value)
 		return segment.Write(offset, &mv)
 	}
-	hashIndex := offset % CellsPerKeccak
-	if hashIndex < InputCellsPerKeccak {
+	hashIndex := offset % cellsPerKeccak
+	if hashIndex < inputCellsPerKeccak {
 		return errors.New("cannot infer value")
 	}
 
 	startOffset := offset - hashIndex
 	var data [200]byte
-	for i := uint64(0); i < InputCellsPerKeccak; i++ {
+	for i := uint64(0); i < inputCellsPerKeccak; i++ {
 		value := segment.Peek(startOffset + i)
 		if !value.Known() {
 			return fmt.Errorf("cannot infer value: input value at offset %d is unknown", startOffset+i)
@@ -70,12 +72,12 @@ func (k *Keccak) InferValue(segment *memory.Segment, offset uint64) error {
 		binary.LittleEndian.PutUint64(output[i*8:i*8+8], dataU64[i])
 	}
 
-	for i := 0; i < InputCellsPerKeccak; i++ {
+	for i := 0; i < inputCellsPerKeccak; i++ {
 		var bytes [32]byte
 		copy(bytes[:], output[i*25:i*25+25])
 		//This is 25*8 bits which is smaller than max felt 252 bits so no need to check the error
 		v, _ := fp.LittleEndian.Element(&bytes)
-		k.cache[startOffset+InputCellsPerKeccak+uint64(i)] = v
+		k.cache[startOffset+inputCellsPerKeccak+uint64(i)] = v
 	}
 	value = k.cache[offset]
 	mv := memory.MemoryValueFromFieldElement(&value)
@@ -87,5 +89,70 @@ func (k *Keccak) String() string {
 }
 
 func (k *Keccak) GetAllocatedSize(segmentUsedSize uint64, vmCurrentStep uint64) (uint64, error) {
-	return getBuiltinAllocatedSize(segmentUsedSize, vmCurrentStep, k.ratio, InputCellsPerKeccak, instancesPerComponentKeccak, CellsPerKeccak)
+	return getBuiltinAllocatedSize(segmentUsedSize, vmCurrentStep, k.ratio, inputCellsPerKeccak, instancesPerComponentKeccak, cellsPerKeccak)
+}
+
+type AirPrivateBuiltinKeccak struct {
+	Index   int    `json:"index"`
+	InputS0 string `json:"input_s0"`
+	InputS1 string `json:"input_s1"`
+	InputS2 string `json:"input_s2"`
+	InputS3 string `json:"input_s3"`
+	InputS4 string `json:"input_s4"`
+	InputS5 string `json:"input_s5"`
+	InputS6 string `json:"input_s6"`
+	InputS7 string `json:"input_s7"`
+}
+
+func (k *Keccak) GetAirPrivateInput(keccakSegment *memory.Segment) []AirPrivateBuiltinKeccak {
+	valueMapping := make(map[int]AirPrivateBuiltinKeccak)
+	for index, value := range keccakSegment.Data {
+		if !value.Known() {
+			continue
+		}
+		idx, stateIndex := index/cellsPerKeccak, index%cellsPerKeccak
+		if stateIndex >= inputCellsPerKeccak {
+			continue
+		}
+
+		builtinValue, exists := valueMapping[idx]
+		if !exists {
+			builtinValue = AirPrivateBuiltinKeccak{Index: idx}
+		}
+
+		valueBig := big.Int{}
+		value.Felt.BigInt(&valueBig)
+		valueHex := fmt.Sprintf("0x%x", &valueBig)
+		if stateIndex == 0 {
+			builtinValue.InputS0 = valueHex
+		} else if stateIndex == 1 {
+			builtinValue.InputS1 = valueHex
+		} else if stateIndex == 2 {
+			builtinValue.InputS2 = valueHex
+		} else if stateIndex == 3 {
+			builtinValue.InputS3 = valueHex
+		} else if stateIndex == 4 {
+			builtinValue.InputS4 = valueHex
+		} else if stateIndex == 5 {
+			builtinValue.InputS5 = valueHex
+		} else if stateIndex == 6 {
+			builtinValue.InputS6 = valueHex
+		} else if stateIndex == 7 {
+			builtinValue.InputS7 = valueHex
+		}
+		valueMapping[idx] = builtinValue
+	}
+
+	values := make([]AirPrivateBuiltinKeccak, 0)
+
+	sortedIndexes := make([]int, 0, len(valueMapping))
+	for index := range valueMapping {
+		sortedIndexes = append(sortedIndexes, index)
+	}
+	sort.Ints(sortedIndexes)
+	for _, index := range sortedIndexes {
+		value := valueMapping[index]
+		values = append(values, value)
+	}
+	return values
 }
